+++ src/js/utils/lottoValidation.js
