+++ src/js/utils/utils.js
@@ -1,5 +1,5 @@
 import { LOTTO_NUMBERS } from './constants.js';
 
-export default function getRandomNumber() {
+export function getRandomNumber() {
   return Math.floor(Math.random() * LOTTO_NUMBERS.LOTTO_MAX_NUM) + 1;
 }
